+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 2.1.24-rc.0
 - added Minter
+- update for pluasible library
 
 ## 2.1.23
 - added dispenser redirect