+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - added Minter
 - update for pluasible library
 - added Create Links button ffor Collections
+- added Mnemonic network support
 
 ## 2.1.23
 - added dispenser redirect